dds/requirements.txt
# Minecraft Farm Bot Dependencies
# Made by DDS
pyautogui==0.9.54
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
keyboard==0.13.5
mouse==0.7.1
colorama==0.4.6
configparser==6.0.0
threading
time
logging 